Update README.md
The chatbot is not an application based but a whole website based program. It gives users an automated way to communicate with the government officials. Chatbot saves time by automatically answering the customer questions and providing accurate responses. It is a free to use application i.e. you don't need to enroll or give any fee for asking or resolving your query. It explains all the acts and regulations related to the mining industry. It has a voice assistant and is available with multilingual support. If the user does not feel satisfied with the answer then the user can have personalized assistance. It has a special feature of emergency support.